requirements.txt
requests==2.32.3
requests-oauthlib>=1.3.1
Pillow==10.3.0
python-dotenv==1.0.1
fastapi==0.112.2
uvicorn==0.30.6
tenacity==8.5.0
httpx==0.27.0